refactor
- split User, Role, Privileges into separate Views, ViewModels,
Repositories
