+++ src/main/java/atbmhttt/atbmcq_16/Router.java
@@ -21,7 +21,7 @@ public class Router {
     }
 
     public static void navigateToAdminDashboard(String username, String password) {
-        AdminView view = new AdminView(username, password);
+        AdminView view = new AdminView();
         view.start(P_STAGE);
     }
