+++ src/main/java/atbmhttt/atbmcq_16/admin/Repositories/PrivilegesRepository.java
@@ -1,0 +1,5 @@
+package atbmhttt.atbmcq_16.admin.Repositories;
+
+public class PrivilegesRepository {
+    // PLACE-HOLDER
+}
