+++ src/main/java/atbmhttt/atbmcq_16/admin/Repositories/RolePrivilegesRepository.java
@@ -1,0 +1,13 @@
+package atbmhttt.atbmcq_16.admin.Repositories;
+
+import java.sql.SQLException;
+import java.util.List;
+
+public class RolePrivilegesRepository {
+    private RolesRepository rolesRepository = new RolesRepository();
+
+    public List<String[]> getPdbRoles() throws SQLException {
+        return rolesRepository.getPdbRoles();
+    }
+
+}
