+++ src/main/java/atbmhttt/atbmcq_16/admin/Repositories/UserPrivilegesRepository.java
@@ -1,0 +1,12 @@
+package atbmhttt.atbmcq_16.admin.Repositories;
+
+import java.sql.SQLException;
+import java.util.List;
+
+public class UserPrivilegesRepository {
+    private UsersRepository usersRepository = new UsersRepository();
+
+    public List<String[]> getUsersWithDetails() throws SQLException {
+        return usersRepository.getUsersWithDetails();
+    }
+}
