+++ src/main/java/atbmhttt/atbmcq_16/admin/Repositories/UsersRepository.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import atbmhttt.atbmcq_16.admin.DatabaseConnection;
+import atbmhttt.atbmcq_16.DatabaseConnection;
 
 public class UsersRepository {
