+++ src/main/java/atbmhttt/atbmcq_16/admin/ViewModels/RolePrivilegesViewModel.java
@@ -1,0 +1,14 @@
+package atbmhttt.atbmcq_16.admin.ViewModels;
+
+import java.sql.SQLException;
+import java.util.List;
+
+import atbmhttt.atbmcq_16.admin.Repositories.RolePrivilegesRepository;
+
+public class RolePrivilegesViewModel {
+    private RolePrivilegesRepository repository = new RolePrivilegesRepository();
+
+    public List<String[]> getPdbRoles() throws SQLException {
+        return repository.getPdbRoles();
+    }
+}
