+++ src/main/java/module-info.java
@@ -1,7 +1,7 @@
 module atbmhttt.atbmcq_16 {
     requires transitive javafx.controls; // Make javafx.controls available to dependent modules
     requires transitive javafx.graphics;
-    requires java.sql;
+    requires transitive java.sql;
     requires javafx.base;
 
     opens atbmhttt.atbmcq_16 to javafx.fxml;
